refactor: save location cookie to be used only by auth service
